Added a takeoff command to mmap.py
The takeoff command will throttle up via RC override and then hold. Then
a mavlink mission item with a MAV_CMD_NAV_TAKEOFF message is sent (in guide mode)
to allow the quad to take off at the given altitude. Currently the takeoff_alt is
set in mmap.py but should be moved to the web front end as a user parameter.